+++ python/GafferUI/VectorDataWidget.py
@@ -42,6 +42,7 @@ import IECore
 
 import Gaffer
 import GafferUI
+from GafferUI.ColorSwatch import _Checker
 from ._TableView import _TableView
 
 from Qt import QtCore
@@ -879,7 +880,8 @@ class _Model( QtCore.QAbstractTableModel ) :
 				if self.__header is not None :
 					result = self.__header[self.columnToDataIndex(section)[0]]
 					if column.accessor.numColumns() > 1 :
-						result += "." + column.accessor.headerLabel( column.relativeColumnIndex )
+						suffix = column.accessor.headerLabel( column.relativeColumnIndex )
+						result += ( "." + suffix ) if suffix is not None else ""
 				else :
 					result = column.accessor.headerLabel( column.relativeColumnIndex )
 				return GafferUI._Variant.toVariant( result )
@@ -1017,10 +1019,7 @@ class _CompoundDataAccessor( _DataAccessor ) :
 
 	def headerLabel( self, columnIndex ) :
 
-		if isinstance( self.data(), ( IECore.Color3fVectorData, IECore.Color4fVectorData ) ) :
-			return [ "R", "G", "B", "A" ][columnIndex]
-		else :
-			return [ "X", "Y", "Z", "W" ][columnIndex]
+		return [ "X", "Y", "Z", "W" ][columnIndex]
 
 	def setElement( self, rowIndex, columnIndex, value ) :
 
@@ -1032,9 +1031,6 @@ class _CompoundDataAccessor( _DataAccessor ) :
 
 		return GafferUI._Variant.toVariant( self.data()[rowIndex][columnIndex] )
 
-_DataAccessor.registerType( IECore.Color3fVectorData.staticTypeId(), _CompoundDataAccessor )
-_DataAccessor.registerType( IECore.Color4fVectorData.staticTypeId(), _CompoundDataAccessor )
-
 _DataAccessor.registerType( IECore.V2iVectorData.staticTypeId(), _CompoundDataAccessor )
 _DataAccessor.registerType( IECore.V2fVectorData.staticTypeId(), _CompoundDataAccessor )
 _DataAccessor.registerType( IECore.V2dVectorData.staticTypeId(), _CompoundDataAccessor )
@@ -1042,6 +1038,57 @@ _DataAccessor.registerType( IECore.V2dVectorData.staticTypeId(), _CompoundDataAc
 _DataAccessor.registerType( IECore.V3iVectorData.staticTypeId(), _CompoundDataAccessor )
 _DataAccessor.registerType( IECore.V3fVectorData.staticTypeId(), _CompoundDataAccessor )
 _DataAccessor.registerType( IECore.V3dVectorData.staticTypeId(), _CompoundDataAccessor )
+
+class _ColorDataAccessor( _CompoundDataAccessor ) :
+
+	def __init__( self, data ) :
+
+		_CompoundDataAccessor.__init__( self, data )
+
+	def setElement( self, rowIndex, columnIndex, value ) :
+
+		if columnIndex < self.numColumns() - 1 :
+			_CompoundDataAccessor.setElement( self, rowIndex, columnIndex, value )
+		else :
+			self.data()[rowIndex] = GafferUI._Variant.fromVariant( value )
+
+	def getElement( self, rowIndex, columnIndex ) :
+
+		if columnIndex < self.numColumns() - 1 :
+			return _CompoundDataAccessor.getElement( self, rowIndex, columnIndex )
+		else :
+			return GafferUI._Variant.toVariant( self.data()[rowIndex] )
+
+class _Color3fDataAccessor( _ColorDataAccessor ) :
+
+	def __init__( self, data ) :
+
+		_ColorDataAccessor.__init__( self, data )
+
+	def numColumns( self ) :
+
+		return 4
+
+	def headerLabel( self, columnIndex ) :
+
+		return [ "R", "G", "B", None ][columnIndex]
+
+class _Color4fDataAccessor( _ColorDataAccessor ) :
+
+	def __init__( self, data ) :
+
+		_ColorDataAccessor.__init__( self, data )
+
+	def numColumns( self ) :
+
+		return 5
+
+	def headerLabel( self, columnIndex ) :
+
+		return [ "R", "G", "B", "A", None ][columnIndex]
+
+_DataAccessor.registerType( IECore.Color3fVectorData.staticTypeId(), _Color3fDataAccessor )
+_DataAccessor.registerType( IECore.Color4fVectorData.staticTypeId(), _Color4fDataAccessor )
 
 class _QuatDataAccessor( _CompoundDataAccessor ) :
 
@@ -1215,10 +1262,19 @@ class _Delegate( QtWidgets.QStyledItemDelegate ) :
 		# fall through to the base class which will provide a default
 		# editor.
 		if self.__editor is not None :
-			self.__editor._qtWidget().setParent( parent )
+			if not isinstance( self.__editor, GafferUI.Window ) :
+				self.__editor._qtWidget().setParent( parent )
 			return self.__editor._qtWidget()
 		else :
 			return QtWidgets.QStyledItemDelegate.createEditor( self, parent, option, index )
+
+	def updateEditorGeometry( self, editor, option, index ) :
+
+		if not isinstance( self.__editor, GafferUI.Window ) :
+			QtWidgets.QStyledItemDelegate.updateEditorGeometry( self, editor, option, index )
+		else :
+			# Windows can't be conformed to the cell geometry
+			pass
 
 	def setEditorData( self, editor, index ) :
 
@@ -1311,8 +1367,6 @@ _Delegate.registerType( IECore.UIntVectorData.staticTypeId(), _NumericDelegate )
 _Delegate.registerType( IECore.Int64VectorData.staticTypeId(), _NumericDelegate )
 _Delegate.registerType( IECore.UInt64VectorData.staticTypeId(), _NumericDelegate )
 _Delegate.registerType( IECore.FloatVectorData.staticTypeId(), _NumericDelegate )
-_Delegate.registerType( IECore.Color3fVectorData.staticTypeId(), _NumericDelegate )
-_Delegate.registerType( IECore.Color4fVectorData.staticTypeId(), _NumericDelegate )
 
 _Delegate.registerType( IECore.V2iVectorData.staticTypeId(), _NumericDelegate )
 _Delegate.registerType( IECore.V2fVectorData.staticTypeId(), _NumericDelegate )
@@ -1337,6 +1391,110 @@ _Delegate.registerType( IECore.Box3dVectorData.staticTypeId(), _NumericDelegate 
 
 _Delegate.registerType( IECore.QuatfVectorData.staticTypeId(), _NumericDelegate )
 _Delegate.registerType( IECore.QuatdVectorData.staticTypeId(), _NumericDelegate )
+
+# A delegate that creates a `_NumericDelegate` for each of the color components
+# and an editable color swatch for the color column.
+class _ColorDelegate( _Delegate ) :
+
+	def __init__( self ) :
+
+		_Delegate.__init__( self )
+
+		self.__colorChooser = None
+		self.__popup = None
+
+		self.__checkerBoardColor0 = imath.Color3f( 0.1 )
+		self.__checkerBoardColor1 = imath.Color3f( 0.2 )
+
+	def paint( self, painter, option, index ) :
+
+		value = self.__colorData( index )
+
+		if isinstance( value, float ) :
+			_Delegate.paint( self, painter, option, index )
+		else :
+			# in PyQt, option is passed to us correctly as a QStyleOptionViewItemV4,
+			# but in PySide it is merely a QStyleOptionViewItem and we must "cast" it.
+			if hasattr( QtGui, "QStyleOptionViewItemV4" ) :
+				option = QtGui.QStyleOptionViewItemV4( option )
+
+			# in PyQt, we can access the widget with option.widget, but in PySide it
+			# is always None for some reason, so we jump through some hoops to get the
+			# widget via our parent.
+			widget = QtCore.QObject.parent( self.parent() )
+
+			# draw the background
+
+			widget.style().drawControl( QtWidgets.QStyle.CE_ItemViewItem, option, painter, widget )
+			displayTransform = GafferUI.DisplayTransform.get()
+			transformedColor = displayTransform( value )
+
+			opaqueCheckerColor0 = GafferUI.Widget._qtColor( transformedColor )
+			opaqueCheckerColor1 = opaqueCheckerColor0
+			transparentCheckerColor0 = opaqueCheckerColor0
+			transparentCheckerColor1 = opaqueCheckerColor0
+
+			if isinstance( value, imath.Color4f ) :
+				transparentCheckerColor0 = self.__checkerBoardColor0 * ( 1.0 - value.a ) + imath.Color3f( value.r, value.g, value.b ) * value.a
+				transparentCheckerColor1 = self.__checkerBoardColor1 * ( 1.0 - value.a ) + imath.Color3f( value.r, value.g, value.b ) * value.a
+
+				transparentCheckerColor0 = GafferUI.Widget._qtColor( displayTransform( transparentCheckerColor0 ) )
+				transparentCheckerColor1 = GafferUI.Widget._qtColor( displayTransform( transparentCheckerColor1 ) )
+
+			padding = 2
+			size = imath.V2i( min( option.rect.height() - padding * 2, option.rect.width() ), ( option.rect.height() - padding * 2 ) * 0.5 )
+			topLeft = imath.V2i( option.rect.x() + option.rect.width() * 0.5 - size.x * 0.5, option.rect.y() + padding )
+
+			topRect = QtCore.QRectF( topLeft.x, topLeft.y, size.x, size.y )
+			bottomRect = QtCore.QRectF( topLeft.x, topLeft.y + size.y, size.x, size.y )
+			_Checker._paintRectangle(
+				painter,
+				topRect,
+				opaqueCheckerColor0,
+				opaqueCheckerColor1
+			)
+			_Checker._paintRectangle(
+				painter,
+				bottomRect,
+				transparentCheckerColor0,
+				transparentCheckerColor1
+			)
+
+	def _createEditorInternal( self, index ) :
+
+		value = self.__colorData( index )
+
+		if isinstance( value, float ) :
+			return GafferUI.NumericWidget( value )
+		else :
+			self.__colorChooser = GafferUI.ColorChooser( value )
+			self.__popup = GafferUI.PopupWindow( "", child = self.__colorChooser )
+
+			self.__popup.popup()
+
+			return self.__popup
+
+	def setEditorData( self, editor, index ) :
+
+		value = self.__colorData( index )
+		if isinstance( value, float ) :
+			_Delegate.setEditorData( self, editor, index )
+		else :
+			self.__colorChooser.setColor( value )
+
+	def setModelData( self, editor, model, index ) :
+
+		if self.__colorChooser is None :
+			_Delegate.setModelData( self, editor, model, index )
+		else :
+			model.setData( index, self.__colorChooser.getColor(), QtCore.Qt.EditRole )
+
+	def __colorData( self, index ) :
+
+		return index.model().data( index, QtCore.Qt.DisplayRole )
+
+_Delegate.registerType( IECore.Color3fVectorData.staticTypeId(), _ColorDelegate )
+_Delegate.registerType( IECore.Color4fVectorData.staticTypeId(), _ColorDelegate )
 
 class _BoolDelegate( _Delegate ) :